+++ app/api/v2/models/meals.py
@@ -9,25 +9,12 @@ class MealModel:
         self.meal_id = data.get('meal_id')
         self.price = data.get('price')
         self.table = data.get('table_name')
-        self.db = create_conn()
-      
-    def get_by_id(self):
-        con, response = self.db, None
-        cur = con.cursor(cursor_factory=RealDictCursor)
-        try:
-            cur.execute("""select * from meals WHERE meal_id='{}'
-                """.format(self.meal_id))
-            response = cur.fetchone()
-        except psycopg2.DatabaseError as e:
-            return {'message': '{}'.format(e)}
-        con.close()
-        return response
+        self.db = create_conn()      
 
     def get_by_name(self):
         con, response = self.db, None
         cur = con.cursor()
         try:
-            
             cur.execute("""select * from meals WHERE meal_name='{}'
             """.format(self.meal_name))
             response = cur.fetchall()
@@ -36,50 +23,6 @@ class MealModel:
         con.close()
         return response
 
-    def update_name(self):
-        con, response = self.db, None
-        cur = con.cursor(cursor_factory=RealDictCursor)
-        try:
-            cur.execute("""UPDATE meals SET meal_name='{}', price='{}' WHERE meal_id='{}'
-            """.format(self.meal_name, self.price, self.meal_id))
-            con.commit()
-        except psycopg2.DatabaseError as e:
-            return {'message': '{}'.format(e)}
-        con.close()
+  
 
-    def get_all(self):
-        con, response = self.db, None
-        cur = con.cursor(cursor_factory=RealDictCursor)
-        try:
-            cur.execute("select * from {}".format(self.table))
-            response = cur.fetchall()
-        except psycopg2.DatabaseError as e:
-            return {'message': '{}'.format(e)}
-        con.close()
-        return response
-
-    def delete(self):
-        con = self.db
-        cur = con.cursor(cursor_factory=RealDictCursor)
-        try:
-            cur.execute("""delete from meals where meal_id='{}'
-            """.format(self.meal_id))
-            con.commit()
-        except psycopg2.DatabaseError as e:
-            return {'message': '{}'.format(e)}
-        con.close()
-
-    def save(self):
-        data, conn = None, self.db
-        cursor = conn.cursor(cursor_factory=RealDictCursor)
-        try:
-            query = """insert into meals (meal_name, price) values('{}','{}')
-            """.format(self.meal_name, self.price)
-            cursor.execute(query)
-            conn.commit()
-            data = cursor.fetchone()
-            cursor.close()   
-        except psycopg2.DatabaseError as e:
-            return {'message': '{}'.format(e)}
-        return data
